# docs/swagger.yaml
basePath: /
definitions:
  challenge-backend-1_internal_model.AddressResponse:
    properties:
      city:
        type: string
      country:
        type: string
      created_at:
        type: integer
      id:
        type: string
      postal_code:
        type: string
      province:
        type: string
      street:
        type: string
      updated_at:
        type: integer
    type: object
  challenge-backend-1_internal_model.ContactResponse:
    properties:
      addresses:
        items:
          $ref: '#/definitions/challenge-backend-1_internal_model.AddressResponse'
        type: array
      created_at:
        type: integer
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
      updated_at:
        type: integer
    type: object
  challenge-backend-1_internal_model.CreateAddressRequest:
    properties:
      city:
        maxLength: 255
        type: string
      country:
        maxLength: 100
        type: string
      postal_code:
        maxLength: 10
        type: string
      province:
        maxLength: 255
        type: string
      street:
        maxLength: 255
        type: string
    type: object
  challenge-backend-1_internal_model.CreateContactRequest:
    properties:
      email:
        maxLength: 200
        type: string
      first_name:
        maxLength: 100
        type: string
      last_name:
        maxLength: 100
        type: string
      phone:
        maxLength: 20
        type: string
    required:
    - first_name
    type: object
  challenge-backend-1_internal_model.ErrorResponse:
    properties:
      errors:
        type: string
    type: object
  challenge-backend-1_internal_model.LoginUserRequest:
    properties:
      id:
        maxLength: 100
        type: string
      password:
        maxLength: 100
        type: string
    required:
    - id
    - password
    type: object
  challenge-backend-1_internal_model.RegisterUserRequest:
    properties:
      id:
        maxLength: 100
        type: string
      name:
        maxLength: 100
        type: string
      password:
        maxLength: 100
        type: string
    required:
    - id
    - name
    - password
    type: object
  challenge-backend-1_internal_model.UpdateAddressRequest:
    properties:
      city:
        maxLength: 255
        type: string
      country:
        maxLength: 100
        type: string
      postal_code:
        maxLength: 10
        type: string
      province:
        maxLength: 255
        type: string
      street:
        maxLength: 255
        type: string
    type: object
  challenge-backend-1_internal_model.UpdateContactRequest:
    properties:
      email:
        maxLength: 200
        type: string
      first_name:
        maxLength: 100
        type: string
      last_name:
        maxLength: 100
        type: string
      phone:
        maxLength: 20
        type: string
    required:
    - first_name
    type: object
  challenge-backend-1_internal_model.UpdateUserRequest:
    properties:
      name:
        maxLength: 100
        type: string
      password:
        maxLength: 100
        type: string
    type: object
  challenge-backend-1_internal_model.UserResponse:
    properties:
      created_at:
        type: integer
      id:
        type: string
      name:
        type: string
      token:
        type: string
      updated_at:
        type: integer
    type: object
  challenge-backend-1_internal_model.WebResponse-array_challenge-backend-1_internal_model_AddressResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/challenge-backend-1_internal_model.AddressResponse'
        type: array
    type: object
  challenge-backend-1_internal_model.WebResponse-array_challenge-backend-1_internal_model_ContactResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/challenge-backend-1_internal_model.ContactResponse'
        type: array
    type: object
  challenge-backend-1_internal_model.WebResponse-bool:
    properties:
      data:
        type: boolean
    type: object
  challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_AddressResponse:
    properties:
      data:
        $ref: '#/definitions/challenge-backend-1_internal_model.AddressResponse'
    type: object
  challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_ContactResponse:
    properties:
      data:
        $ref: '#/definitions/challenge-backend-1_internal_model.ContactResponse'
    type: object
  challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_UserResponse:
    properties:
      data:
        $ref: '#/definitions/challenge-backend-1_internal_model.UserResponse'
    type: object
host: localhost:8080
info:
  contact: {}
  description: Challenge Backend 1
  title: Challenge Backend 1
  version: 1.0.0
paths:
  /api/contacts:
    get:
      consumes:
      - application/json
      description: List contacts
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: Email
        in: query
        name: email
        type: string
      - description: Phone
        in: query
        name: phone
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-array_challenge-backend-1_internal_model_ContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List contacts
      tags:
      - Contact API
    post:
      consumes:
      - application/json
      description: Create new contact
      parameters:
      - description: Create Contact Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/challenge-backend-1_internal_model.CreateContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_ContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create new contact
      tags:
      - Contact API
  /api/contacts/{contactId}:
    delete:
      consumes:
      - application/json
      description: Delete contact
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-bool'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete contact
      tags:
      - Contact API
    get:
      consumes:
      - application/json
      description: Get contact
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_ContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get contact
      tags:
      - Contact API
    put:
      consumes:
      - application/json
      description: Update contact
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      - description: Update Contact Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/challenge-backend-1_internal_model.UpdateContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_ContactResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update contact
      tags:
      - Contact API
  /api/contacts/{contactId}/addresses:
    get:
      consumes:
      - application/json
      description: List addresses
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-array_challenge-backend-1_internal_model_AddressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List addresses
      tags:
      - Address API
    post:
      consumes:
      - application/json
      description: Create new address
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      - description: Create Address Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/challenge-backend-1_internal_model.CreateAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_AddressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create new address
      tags:
      - Address API
  /api/contacts/{contactId}/addresses/{addressId}:
    delete:
      consumes:
      - application/json
      description: Delete address
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      - description: Address ID
        in: path
        name: addressId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-bool'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete address
      tags:
      - Address API
    get:
      consumes:
      - application/json
      description: Get address
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      - description: Address ID
        in: path
        name: addressId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_AddressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get address
      tags:
      - Address API
    put:
      consumes:
      - application/json
      description: Update address
      parameters:
      - description: Contact ID
        in: path
        name: contactId
        required: true
        type: string
      - description: Address ID
        in: path
        name: addressId
        required: true
        type: string
      - description: Update Address Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/challenge-backend-1_internal_model.UpdateAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_AddressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update address
      tags:
      - Address API
  /api/users:
    delete:
      consumes:
      - application/json
      description: Logout user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-bool'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - User API
    post:
      consumes:
      - application/json
      description: Register new user
      parameters:
      - description: Register User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/challenge-backend-1_internal_model.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      summary: Register new user
      tags:
      - User API
  /api/users/_current:
    get:
      consumes:
      - application/json
      description: Get current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get current user
      tags:
      - User API
    patch:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: Update User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/challenge-backend-1_internal_model.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - User API
  /api/users/_login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: Login User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/challenge-backend-1_internal_model.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.WebResponse-challenge-backend-1_internal_model_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/challenge-backend-1_internal_model.ErrorResponse'
      summary: Login user
      tags:
      - User API
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"